arb-finder/pyproject.toml
[tool.poetry]
name = "arb-finder"
version = "0.1.0"
description = "Cross-exchange arbitrage opportunity finder"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
websockets = "^12.0"
orjson = "^3.9"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
